Criação de demands.js e reorganização de funções para maior raestrabilidade e reuso do código
